+++ CHANGELOG.md
@@ -7,6 +7,14 @@ We follow [Semantic Versioning](https://semver.org/):
 - **MAJOR** version `X.0.0`: Increment for incompatible API changes or breaking changes.
 - **MINOR** version `0.Y.0`: Increment for backward-compatible additions (new features, enhancements).
 - **PATCH** version `0.0.Z`: Increment for backward-compatible bug fixes and documentation tweaks.
+## [0.4.1] - 2026-02-28
+### Change
+- change floating shape component.
+- Inhance navigation section 
+- inhanched hero section 
+- inhanched theme toggle 
+- inhance overall UI
+
 
 ## [0.4.1] - 2026-02-28